ユーザ登録処理をコントローラに切り出す
